Renamed db folder to src
