+++ nodeapp/app.js
@@ -1,5 +1,5 @@
 // Initialize project
-const db = require('./db/db')
+const db = require('./src/db')
 const path = require('path')
 require('dotenv').config({path: path.join(__dirname, '../.env')})
 const express = require('express');
